Move functions for extracting metric name from labels & matchers into its own package, to prevent circular imports.
