+++ pkg/util/matchers.go
@@ -1,9 +1,6 @@
 package util
 
 import (
-	"fmt"
-
-	"github.com/prometheus/common/model"
 	"github.com/prometheus/prometheus/pkg/labels"
 )
 
@@ -23,37 +20,4 @@ func SplitFiltersAndMatchers(allMatchers []*labels.Matcher) (filters, matchers [
 		}
 	}
 	return
-}
-
-// ExtractMetricNameFromMetric extract the metric name from a model.Metric
-func ExtractMetricNameFromMetric(m model.Metric) (model.LabelValue, error) {
-	for name, value := range m {
-		if name == model.MetricNameLabel {
-			return value, nil
-		}
-	}
-	return "", fmt.Errorf("no MetricNameLabel for chunk")
-}
-
-// ExtractMetricNameMatcherFromMatchers extracts the metric name from a set of matchers
-func ExtractMetricNameMatcherFromMatchers(matchers []*labels.Matcher) (*labels.Matcher, []*labels.Matcher, bool) {
-	// Handle the case where there is no metric name and all matchers have been
-	// filtered out e.g. {foo=""}.
-	if len(matchers) == 0 {
-		return nil, matchers, false
-	}
-
-	outMatchers := make([]*labels.Matcher, len(matchers)-1)
-	for i, matcher := range matchers {
-		if matcher.Name != model.MetricNameLabel {
-			continue
-		}
-
-		// Copy other matchers, excluding the found metric name matcher
-		copy(outMatchers, matchers[:i])
-		copy(outMatchers[i:], matchers[i+1:])
-		return matcher, outMatchers, true
-	}
-	// Return all matchers if none are metric name matchers
-	return nil, matchers, false
 }
